\name{nrsf}
\docType{data}
\alias{nrsf}
\title{nrsf data}
\description{
 This is a subset of the neuron-restrictive silencer factor (NRSF) data
 containing the information of the sequence tags that are uniquely mapped
 (up to two mismatches allowed) to chromosomes 22 and Y of human genome.
}
\usage{data(nrsf)}
\source{Science 316, 1497-1502.}
\references{David S. Johnson, Ali Mortazavi, Richard M. Myers, Barbara
  Wold.  Genome-Wide Mapping of in Vivo Protein-DNA Interactions.
  Science 316, 1497-1502.}
\keyword{datasets}
